Fix ArticleHeader folder structure
