+++ apps/codeforafrica/src/components/ArticleHeader/ArticleHeader.snap.js
@@ -1,0 +1,9 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`<ArticleHeader /> renders unchanged 1`] = `
+<div>
+  <div
+    class="MuiBox-root css-j7qwjs"
+  />
+</div>
+`;
