--- apps/codeforafrica/src/components/ArticleHeader/index.snap.js
@@ -1,13 +1,0 @@
-// Jest Snapshot v1, https://goo.gl/fbAQLP
-
-exports[`<ArticleHeader /> renders unchanged 1`] = `
-<div>
-  <div
-    class="MuiGrid-root MuiGrid-container css-11lq3yg-MuiGrid-root"
-  >
-    <div
-      class="MuiGrid-root MuiGrid-item MuiGrid-grid-xs-12 css-18qart5-MuiGrid-root"
-    />
-  </div>
-</div>
-`;
